# Conflicts:
#	data/testdb.mv.db
#	data/testdb.trace.db
#	src/main/java/org/example/Main.java

+++ src/main/java/org/example/api/service/CustomerService.java
@@ -1,5 +1,6 @@
 package org.example.api.service;
 
+import org.example.api.data.entity.Account;
 import org.example.api.data.entity.Customer;
 import org.example.api.data.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,26 +12,25 @@ import java.util.Optional;
 @Service
 public class CustomerService {
 
-    @Autowired
-    private CustomerRepository customerRepository;
+  @Autowired private CustomerRepository customerRepository;
 
-    public List<Customer> findAll() {
-        return customerRepository.findAll();
-    }
+  public List<Customer> findAll() {
+    return customerRepository.findAll();
+  }
 
-    public Optional<Customer> findById(Long customerId) {
-        return customerRepository.findById(customerId);
-    }
+  public Optional<Customer> findById(Integer customerId) {
+    return customerRepository.findById(customerId);
+  }
 
-    public Customer save(Customer customer) {
-        return customerRepository.save(customer);
-    }
+  public Customer save(Customer customer) {
+    return customerRepository.save(customer);
+  }
 
-    public void deleteById(Long customerId) {
-        customerRepository.deleteById(customerId);
-    }
+  public void deleteById(Integer customerId) {
+    customerRepository.deleteById(customerId);
+  }
 
-    public Optional<Customer> findByEmail(String email) {
-        return customerRepository.findByEmail(email);
-    }
+  public Optional<Customer> findByEmail(String email) {
+    return customerRepository.findByEmail(email);
+  }
 }
